# app/cache/test/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "Trailburning Journey API",
        "contact": {
            "name": "Matt Allbuery",
            "url": "http://www.trailburning.com/",
            "email": "matt@trailburning.com"
        },
        "version": "2.0"
    },
    "basePath": "/v2",
    "consumes": [
        "application/json",
        "multipart/form-data",
        "application/x-www-form-urlencoded"
    ],
    "produces": [
        "application/json"
    ],
    "paths": {
...
        "/journeys/{id}/events": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Find events by journey",
                "description": "Returns all events of a journey.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the journey",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Event"
                        }
                    },
                    "404": {
                        "description": "Journey not found"
                    }
                }
            },
            "post": {
                "tags": [
                    "Events"
                ],
                "summary": "Add an event",
                "description": "Adds an event.",
                "consumes": [
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the journey the event belongs to",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "The name of the event",
                        "type": "string"
                    },
                    {
                        "name": "about",
                        "in": "formData",
                        "description": "About the event",
                        "type": "string"
                    },
                    {
                        "name": "coords",
                        "in": "formData",
                        "description": "The GPS coordinates of the event in the format '(LNG, LAT)'",
                        "type": "string"
                    },
                    {
                        "name": "position",
                        "in": "formData",
                        "description": "The sort position, is handled automatically if not specified",
                        "type": "integer"
                    },
                    {
                        "name": "custom[0][key]",
                        "in": "formData",
                        "description": "The name of the event",
                        "type": "string"
                    },
                    {
                        "name": "custom[0][value]",
                        "in": "formData",
                        "description": "The name of the event",
                        "type": "string"
                    },
                    {
                        "name": "custom[1][key]",
                        "in": "formData",
                        "description": "The name of the event",
                        "type": "string"
                    },
                    {
                        "name": "custom[1][value]",
                        "in": "formData",
                        "description": "The name of the event",
                        "type": "string"
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Successful operation. The Location header contains a link to the new event.",
                        "headers": {
                            "location": {
                                "description": "Link to the new event.",
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid data."
                    }
                }
            }
        },
        "/events/{id}": {
            "get": {
                "tags": [
                    "Events"
                ],
                "summary": "Find an event by ID",
                "description": "Returns a single event.",
                "produces": [
                    "application/json"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the event to return",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Event"
                        }
                    },
                    "404": {
                        "description": "Event not found"
                    }
                }
            },
            "put": {
                "tags": [
                    "Events"
                ],
                "summary": "Update an event",
                "description": "Updates a event.",
                "consumes": [
                    "application/json",
                    "application/x-www-form-urlencoded"
                ],
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the event to update",
                        "required": true,
                        "type": "string"
                    },
                    {
                        "name": "name",
                        "in": "formData",
                        "description": "The name of the event",
                        "type": "string"
                    },
                    {
                        "name": "about",
                        "in": "formData",
                        "description": "About the event",
                        "type": "string"
                    },
                    {
                        "name": "position",
                        "in": "formData",
                        "description": "The sort position, is handled automatically if not specified",
                        "type": "integer"
                    },
                    {
                        "name": "coords",
                        "in": "formData",
                        "description": "The GPS coordinates of the event",
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "400": {
                        "description": "Invalid data."
                    }
                }
            },
            "delete": {
                "tags": [
                    "Events"
                ],
                "summary": "Delete an event",
                "description": "Deletes the event.",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the event",
                        "required": true,
                        "type": "string"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Successful operation"
                    },
                    "404": {
                        "description": "Journey not found"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Asset": {
            "required": [
                "id",
                "about",
                "category"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "about": {
                    "type": "string"
                },
                "media": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Media"
                    }
                },
                "category": {
                    "$ref": "#/definitions/AssetCategory"
                },
                "credit": {
                    "type": "string"
                }
            },
            "xml": {
                "name": "Asset"
            }
        },
        "AssetCategory": {
            "required": [
                "name",
                "label"
            ],
            "properties": {
                "name": {
                    "type": "string"
                },
                "label": {
                    "type": "string"
                }
            },
            "xml": {
                "name": "AssetCategory"
            }
        },
        "Event": {
            "required": [
                "id",
                "name",
                "about",
                "coords"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "about": {
                    "type": "string"
                },
                "coords": {
                    "description": "point",
                    "type": "Array"
                },
                "assets": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/Asset"
                    }
                },
                "custom": {}
            },
            "xml": {
                "name": "Event"
            }
        },
...
        "Media": {
            "required": [
                "id",
                "mimeType",
                "path"
            ],
            "properties": {
                "id": {
                    "type": "string"
                },
                "path": {
                    "type": "string"
                },
                "mimeType": {
                    "type": "string"
                },
                "metadata": {}
            },
            "xml": {
                "name": "Media"
            }
        },
...
    }
}